Move LogMsg encoding/decoding into re_transport
