+++ crates/re_data_store/Cargo.toml
@@ -47,7 +47,7 @@ puffin.workspace = true
 criterion = "0.4"
 mimalloc.workspace = true
 rand = "0.8"
-re_log_types = { workspace = true, features = ["load", "save"] }
+re_transport = { workspace = true, features = ["decoder", "encoder"] }
 
 [lib]
 bench = false
