+++ crates/re_sdk/Cargo.toml
@@ -35,12 +35,12 @@ image = ["re_log_types/image"]
 [dependencies]
 re_build_info.workspace = true
 re_error.workspace = true
-re_log_types = { workspace = true, features = ["save"] }
+re_log_types.workspace = true
 re_log.workspace = true
 re_memory.workspace = true
 re_sdk_comms = { workspace = true, features = ["client"] }
 re_smart_channel.workspace = true
-re_transport.workspace = true
+re_transport = { workspace = true, features = ["encoder"] }
 
 anyhow.workspace = true
 document-features = "0.2"
