+++ crates/re_transport/src/lib.rs
@@ -1,9 +1,42 @@
 //! Crate that handles transporting of rerun log types.
 
-pub mod stream_rrd_from_http;
+#[cfg(feature = "decoder")]
+pub mod decoder;
+#[cfg(feature = "encoder")]
+#[cfg(not(target_arch = "wasm32"))] // we do no yet support encoding LogMsgs in the browser
+pub mod encoder;
 
+#[cfg(feature = "encoder")]
 #[cfg(not(target_arch = "wasm32"))]
 mod file_sink;
 
+#[cfg(feature = "decoder")]
+pub mod stream_rrd_from_http;
+
+// ---------------------------------------------------------------------
+
+#[cfg(feature = "encoder")]
 #[cfg(not(target_arch = "wasm32"))]
 pub use file_sink::{FileSink, FileSinkError};
+
+// ---------------------------------------------------------------------------
+
+/// Profiling macro for feature "puffin"
+#[doc(hidden)]
+#[macro_export]
+macro_rules! profile_function {
+    ($($arg: tt)*) => {
+        #[cfg(not(target_arch = "wasm32"))]
+        puffin::profile_function!($($arg)*);
+    };
+}
+
+/// Profiling macro for feature "puffin"
+#[doc(hidden)]
+#[macro_export]
+macro_rules! profile_scope {
+    ($($arg: tt)*) => {
+        #[cfg(not(target_arch = "wasm32"))]
+        puffin::profile_scope!($($arg)*);
+    };
+}
