+++ crates/re_viewer/Cargo.toml
@@ -41,20 +41,15 @@ re_build_info.workspace = true
 re_data_store = { workspace = true, features = ["serde"] }
 re_error.workspace = true
 re_format.workspace = true
+re_log_types = { workspace = true, features = ["ecolor", "glam", "image"] }
 re_log.workspace = true
-re_log_types = { workspace = true, features = [
-  "ecolor",
-  "glam",
-  "image",
-  "save",
-  "load",
-] }
 re_memory.workspace = true
 re_query.workspace = true
 re_renderer = { workspace = true, features = ["arrow", "serde"] }
 re_smart_channel.workspace = true
 re_string_interner.workspace = true
 re_tensor_ops.workspace = true
+re_transport = { workspace = true, features = ["decoder", "encoder"] }
 re_ui.workspace = true
 re_ws_comms = { workspace = true, features = ["client"] }
