+++ crates/re_viewer/src/app.rs
@@ -1793,7 +1793,7 @@ fn save_database_to_file(
         let file = std::fs::File::create(path.as_path())
             .with_context(|| format!("Failed to create file at {path:?}"))?;
 
-        re_log_types::encoding::encode(msgs.iter(), file)
+        re_transport::encoder::encode(msgs.iter(), file)
             .map(|_| path)
             .context("Message encode")
     }
@@ -1803,7 +1803,7 @@ fn save_database_to_file(
 fn load_rrd_to_log_db(mut read: impl std::io::Read) -> anyhow::Result<LogDb> {
     crate::profile_function!();
 
-    let decoder = re_log_types::encoding::Decoder::new(read)?;
+    let decoder = re_transport::decoder::Decoder::new(read)?;
 
     let mut log_db = LogDb::default();
     for msg in decoder {
