+++ crates/rerun/Cargo.toml
@@ -67,11 +67,11 @@ web_viewer = [
 re_build_info.workspace = true
 re_error.workspace = true
 re_format.workspace = true
-re_log_types = { workspace = true, features = ["load"] }
+re_log_types.workspace = true
 re_log.workspace = true
 re_memory.workspace = true
 re_smart_channel.workspace = true
-re_transport.workspace = true
+re_transport = { workspace = true, features = ["decoder", "encoder"] }
 re_ws_comms = { workspace = true, features = ["client"] }
 
 anyhow.workspace = true
